requirements.txt
asn1crypto==0.22.0
astroid==1.5.3
autopep8==1.3.2
bcrypt==3.1.3
bleach==2.0.0
cffi==1.10.0
cryptography==1.9
decorator==4.0.11
entrypoints==0.2.3
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.15
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
pandocfilters==1.4.1
paramiko==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psutil==5.2.2
psycopg2==2.7.1
ptyprocess==0.5.2
pyasn1==0.2.3
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pydocstyle==2.0.0
Pygments==2.2.0
pylint==1.7.2
PyNaCl==1.1.2
python-dateutil==2.6.1
pyzmq==16.0.2
qtconsole==4.3.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
wrapt==1.10.10
xmltodict==0.11.0
